@@ -62,32 +42,55 @@
     "flv",
     "dash"
   ],
-  "lint-staged": {
-    "linters": {
-      "packages/**/*.js": [
-        "eslint -c .eslintrc.js --fix",
-        "git add"
-      ]
-    },
-    "ignore": [
-      "**/browser/*.js",
-      "**/dist/*.js"
-    ]
-  },
+  "browserslist": [
+    "> 1%",
+    "last 2 versions",
+    "not dead",
+    "IE 11"
+  ],
   "author": "yinguohui@bytedance.com",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/bytedance/xgplayer/issues"
   },
-  "config": {
-    "commitizen": {
-      "path": "./node_modules/cz-conventional-changelog"
-    }
-  },
   "homepage": "https://github.com/bytedance/xgplayer#readme",
-  "dependencies": {
-    "babel-eslint": "^10.0.3",
-    "http-server": "^0.11.1",
-    "koa-sslify": "^4.0.3"
+  "devDependencies": {
+    "@babel/core": "^7.17.9",
+    "@babel/eslint-parser": "^7.14.3",
+    "@babel/parser": "^7.15.3",
+    "@babel/preset-env": "^7.16.11",
+    "@rollup/plugin-babel": "^5.3.1",
+    "@types/events": "^3.0.0",
+    "@types/jest": "^28.1.1",
+    "@vitejs/plugin-react": "^1.3.2",
+    "add-stream": "^1.0.0",
+    "autoprefixer": "^10.4.0",
+    "babel-jest": "^27.0.6",
+    "brotli-size": "^4.0.0",
+    "cliui": "^7.0.4",
+    "colorette": "^2.0.16",
+    "conventional-changelog": "^3.1.25",
+    "core-js": "3.12.1",
+    "dts-bundle-generator": "^6.4.0",
+    "enquirer": "^2.3.6",
+    "eslint": "^7.28.0",
+    "eslint-plugin-import": "^2.23.4",
+    "eslint-plugin-no-lookahead-lookbehind-regexp": "^0.1.0",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^5.1.0",
+    "eslint-plugin-standard": "^3.1.0",
+    "execa": "^5.1.1",
+    "fs-extra": "^10.0.0",
+    "jest": "^28.1.1",
+    "jest-environment-jsdom": "^28.1.1",
+    "klaw-sync": "^6.0.0",
+    "sade": "^1.7.4",
+    "sass": "^1.43.4",
+    "semver": "^7.3.5",
+    "ts-jest": "^28.0.5",
+    "ts-morph": "^14.0.0",
+    "url-toolkit": "2.1.6",
+    "vite": "^2.9.8",
+    "zlib": "^1.0.5"
   }
 }
