@@ -1,14 +1,17 @@
-import { HlsLivePlayer } from 'xgplayer-hls-live';
-import { HlsVodPlayer } from 'xgplayer-hls-vod';
+/* c8 ignore next 12 */
+import { HlsPlugin } from './plugin'
 
-export class HlsPlayer {
-  constructor (config) {
-    if (config.isLive) {
-      return new HlsLivePlayer(config)
-    } else {
-      return new HlsVodPlayer(config)
-    }
-  }
-}
+/**
+ * @typedef { import ('./hls/buffer-service/decrypt/index').IExternalDecryptor } IExternalDecryptor
+ */
 
-module.exports = HlsPlayer;
+export {
+  ERR,
+  ERR_CODE,
+  StreamingError
+} from 'xgplayer-streaming-shared'
+export * from './hls'
+export { Event as EVENT } from './hls/constants'
+export { HlsPlugin }
+
+export default HlsPlugin
