@@ -43,7 +43,6 @@ class FetchStreamLoader extends BaseLoader {
     }
 
     constructor(seekHandler, config) {
-        console.log('config', config);
         super('fetch-stream-loader');
         this.TAG = 'FetchStreamLoader';
 
@@ -94,12 +93,7 @@ class FetchStreamLoader extends BaseLoader {
             // Safari incorrectly indicates 'no-referrer' as default policy, fuck it
             referrerPolicy: 'no-referrer-when-downgrade'
         };
-        // add additional headers
-        if (typeof this._config.headers === 'object') {
-            for (let key in this._config.headers) {
-                headers.append(key, this._config.headers[key]);
-            }
-        }
+
         // cors is enabled by default
         if (dataSource.cors === false) {
             // no-cors means 'disregard cors policy', which can only be used in ServiceWorker
