@@ -1,89 +1,131 @@
-import Player from 'xgplayer'
+import { BasePlugin, Errors, Events } from 'xgplayer'
 import Flv from './flv/flv'
 
-class FlvJsPlayer extends Player {
-  constructor (options) {
-    super(options)
-    this.flvOpts = { type: 'flv' }
-    this.optionalConfig = {}
-    Player.util.deepCopy(this.flvOpts, options)
-    Player.util.deepCopy(this.optionalConfig, options.flvOptionalConfig)
-    const player = this
-
-    this._start = this.start
-    this.start = () => {
-      if(!window.XgVideoProxy) {
-        this.root.insertBefore(this.video, this.root.firstChild)
-      }
-      setTimeout(() => {
-        this.emit('complete')
-        if(this.danmu && typeof this.danmu.resize === 'function') {
-          this.danmu.resize()
-        }
-      }, 1)
-    }
-
-    Object.defineProperty(player, 'src', {
-      get () {
-        return player.currentSrc
-      },
-      set (url) {
-        player.flv_load(url)
-        let oldVol = player.volume
-        player.video.muted = true
-        Player.util.addClass(player.root, 'xgplayer-is-enter')
-        player.once('playing', function(){
-          Player.util.removeClass(player.root, 'xgplayer-is-enter')
-          player.video.muted = false
-        })
-        player.once('canplay', function () {
-          player.play()
-        })
-      },
-      configurable: true
-    })
-
-    player.once('complete', () => {
-      player.__flv__ = Flv.createPlayer(this.flvOpts, this.optionalConfig)
-      player.emit('flv_complete')
-      player.createInstance(player.__flv__)
-      if(player.config.isLive) {
-        Player.util.addClass(player.root, 'xgplayer-is-live')
-        const live = Player.util.createDom('xg-live', '正在直播', {}, 'xgplayer-live')
-        player.controls.appendChild(live)
-      }
-    })
+class FlvJsPlugin extends BasePlugin {
+  static get isSupported () {
+    return Flv.isSupported
   }
 
-  createInstance (flv) {
-    const player = this
+  static get pluginName () {
+    return 'FlvJsPlugin'
+  }
+
+  static get defaultConfig () {
+    return {
+      mediaDataSource: { type: 'flv' },
+      flvConfig: {}
+    }
+  }
+
+  beforePlayerInit () {
+    if (this.playerConfig.url) {
+      this.flvLoad(this.playerConfig.url)
+    }
+  }
+
+  afterCreate () {
+    const { player } = this
+    this.flv = null
     player.video.addEventListener('contextmenu', function (e) {
       e.preventDefault()
     })
+
+    this.on(Events.URL_CHANGE, (url) => {
+      if (/^blob/.test(url)) {
+        return
+      }
+      player.once(Events.LOADED_DATA, () => {
+        player.play()
+      })
+      this.playerConfig.url = url
+      this.flvLoad(url)
+    })
+    try {
+      BasePlugin.defineGetterOrSetter(player, {
+        url: {
+          get: () => {
+            try {
+              return this.player.video.src
+            } catch (error) {
+              return null
+            }
+          },
+          configurable: true
+        }
+      })
+    } catch (e) {
+      // NOOP
+    }
+  }
+
+  destroy () {
+    const { player } = this
+    this.destroyInstance()
+    BasePlugin.defineGetterOrSetter(player, {
+      url: {
+        get: () => {
+          try {
+            return player.__url
+          } catch (error) {
+            return null
+          }
+        },
+        configurable: true
+      }
+    })
+  }
+
+  destroyInstance () {
+    if (!this.flv) {
+      return
+    }
+    const { player } = this
+    this.flv.unload()
+    this.flv.detachMediaElement(player.video)
+    this.flv.destroy()
+    player.__flv__ = null
+    this.flv = null
+  }
+
+  createInstance (flv) {
+    const { player } = this
+    if (!flv) {
+      return
+    }
+    console.log('createInstance', flv)
     flv.attachMediaElement(player.video)
     flv.load()
     flv.play()
-     
+
     flv.on(Flv.Events.ERROR, (e) => {
-      player.emit('error', new Player.Errors('other', player.config.url))
+      player.emit('error', new Errors('other', player.config.url))
     })
     flv.on(Flv.Events.LOADED_SEI, (timestamp, data) => {
-      player.emit('loaded_sei', timestamp, data);
+      console.log('Flv.Events.LOADED_SEI')
+      player.emit('loaded_sei', timestamp, data)
     })
     flv.on(Flv.Events.STATISTICS_INFO, (data) => {
-      player.emit("statistics_info",data);
+      console.log('Flv.Events.STATISTICS_INFO')
+      player.emit('statistics_info', data)
     })
-    flv.on(Flv.Events.MEDIA_INFO, (data)=>{
-      player.mediainfo = data;
-      player.emit("MEDIA_INFO",data);
-    })
-    player.once('destroy', () => {
-      flv.destroy()
-      player.__flv__ = null
+    flv.on(Flv.Events.MEDIA_INFO, (data) => {
+      player.mediainfo = data
+      player.emit('MEDIA_INFO', data)
+      // console.log('player.autoplay', player.autoplay, player.paused)
+      // if (player.autoplay) {
+      //   player.once('canplay', () => {
+      //     console.log('canplay')
+      //     player.play()
+      //   })
+      // } else if (player.paused) {
+      //   player.pause()
+      // }
     })
   }
-  flv_load (newUrl) {
-    let mediaDataSource = this.flvOpts
+
+  flvLoad (newUrl) {
+    console.log('flvLoad', newUrl)
+    const mediaDataSource = this.config.mediaDataSource
     mediaDataSource.segments = [
       {
         cors: true,
@@ -99,64 +141,43 @@ class FlvJsPlayer extends Player {
     // mediaDataSource.hasVideo = true
     // mediaDataSource.isLive = true
     mediaDataSource.url = newUrl
+    mediaDataSource.isLive = this.playerConfig.isLive
     // mediaDataSource.withCredentials = false
-    this.flv_load_mds(mediaDataSource)
+    this.flvLoadMds(mediaDataSource)
   }
-  flv_load_mds (mediaDataSource) {
-    let player = this
-    if (typeof player.__flv__ !== 'undefined') {
-      if (player.__flv__ != null) {
-        player.__flv__.unload()
-        player.__flv__.detachMediaElement()
-        player.__flv__.destroy()
-        player.__flv__ = null
-      }
-    }
-    player.__flv__ = Flv.createPlayer(mediaDataSource, this.optionalConfig)
 
-    player.__flv__.attachMediaElement(player.video)
-    player.__flv__.load()
+  flvLoadMds (mediaDataSource) {
+    const { player } = this
+    if (typeof this.flv !== 'undefined') {
+      this.destroyInstance()
+    }
+    this.flv = player.__flv__ = Flv.createPlayer(mediaDataSource, this.flvConfig)
+    this.createInstance(this.flv)
+    this.flv.attachMediaElement(player.video)
+    this.flv.load()
   }
 
   switchURL (url) {
-    const player = this
+    const { player, playerConfig } = this
     let curTime = 0
-    if(!player.config.isLive) {
+    if (!playerConfig.isLive) {
       curTime = player.currentTime
     }
-    player.flv_load(url)
-    let oldVol = player.volume
+    player.flvLoad(url)
+    // const oldVol = player.volume
     player.video.muted = true
-    Player.util.addClass(player.root, 'xgplayer-is-enter')
-    player.once('playing', function(){
-      Player.util.removeClass(player.root, 'xgplayer-is-enter')
+    // Util.addClass(player.root, 'xgplayer-is-enter')
+    this.once('playing', function () {
+      // Util.removeClass(player.root, 'xgplayer-is-enter')
       player.video.muted = false
     })
-    player.once('canplay', function () {
-      if(!player.config.isLive) {
+    this.once('canplay', function () {
+      if (!playerConfig.isLive) {
         player.currentTime = curTime
       }
       player.play()
     })
   }
-
-  destroy (isDelDom = true) {
-    let player = this
-    if (player.__flv__) {
-      if (player.__flv__.unload) {
-        player.__flv__.unload()
-      }
-      if (player.__flv__.detachMediaElement) {
-        player.__flv__.detachMediaElement()
-      }
-      if (player.__flv__.destroy) {
-        player.__flv__.destroy()
-      }
-    }
-    player.__flv__ = null
-    super.destroy(isDelDom)  
-  }
 }
-FlvJsPlayer.isSupported = Flv.isSupported
-FlvJsPlayer.FlvJs = Flv
-export default FlvJsPlayer
+
+export default FlvJsPlugin
