@@ -4,8 +4,9 @@ class Buffer {
   constructor () {
     this.buffer = new Uint8Array(0)
   }
+
   write (...buffer) {
-    let self = this
+    const self = this
     buffer.forEach(item => {
       if (item) {
         self.buffer = Concat(Uint8Array, self.buffer, item)
@@ -14,6 +15,7 @@ class Buffer {
       }
     })
   }
+
   static writeUint32 (value) {
     return new Uint8Array([
       value >> 24,
