@@ -1,72 +1,153 @@
-import 'core-js/modules/es6.promise.js'
-import 'core-js/modules/es7.string.pad-start'
-import Player from 'xgplayer'
+import { BasePlugin, Sniffer, Util, Events, Errors } from 'xgplayer'
 import MP4 from './mp4'
 import MSE from './media/mse'
 import Task from './media/task'
 import Buffer from './fmp4/buffer'
 import FMP4 from './fmp4/mp4'
 
-let isEnded = (player, mp4) => {
+let MIME_TYPE = 'audio/mp4; codecs="mp4a.40.5"'
+
+const isEnded = (player, mp4) => {
   if (mp4.meta.endTime - player.currentTime < 2) {
-    let range = player.getBufferedRange()
+    const range = player.getBufferedRange()
     if (player.currentTime - range[1] < 0.1) {
       player.mse.endOfStream()
     }
   }
 }
 
-let errorHandle = (player, err) => {
-  err.vid = player.config.vid
-  err.url = player.src
-  if (err.errd && typeof err.errd === 'object') {
-    if (player.mp4) {
-      err.errd.url = player.mp4.url
-      err.url = player.mp4.url
-      player.mp4.canDownload = false
+const errorHandle = (player, err) => {
+  console.log('errorHandle')
+}
+class M4APlayer extends BasePlugin {
+  static set MIME_TYPE (value) {
+    MIME_TYPE = value
+  }
+
+  static get MIME_TYPE () {
+    return MIME_TYPE
+  }
+
+  static get pluginName () {
+    return 'm4aPlayer'
+  }
+
+  static get defaultConfig () {
+    return {
+      withCredentials: true,
+      videoOnly: false,
+      headers: {
+      },
+      chunkSize: Math.pow(25, 4),
+      preloadTime: 15, // 预加载时长
+      reqTimeLength: 5,
+      segPlay: false,
+      pluginRule: () => {
+        return true
+      }
     }
   }
-  player.emit('DATA_REPORT', err)
-  if (err.errt === 'network' && player.config._backupURL) {
-    player.src = player.config._backupURL
-  } else {
-    player.src = player.config._mainURL
-  }
-  player.switchURL = null
-  player._replay = null
-}
 
-let m4aplayer = function () {
-  let player = this; let sniffer = Player.sniffer; let util = Player.util
-  let Errors = Player.Errors; let mainURL; let backupURL
-  let preloadTime = player.config.preloadTime || 15
-  let waiterTimer
-  Player.m4a = true
-  player.hasEnded = false
-  let list = util.typeOf(player.config.url) === 'Array' ? player.config.url : [{
-    src: player.config.url,
-    name: player.config.name,
-    vid: player.config.vid,
-    poster: player.config.poster
-  }]
-  let url = list[0].src
-  let name = list[0].name
-  let vid = list[0].vid
-  let poster = list[0].poster
-  let rule = player.config.pluginRule || function () { return true }
-  if (!url) {
-    player.emit('error', new Errors('other', player.config.vid))
-    return
+  static isSupported () {
+    return ['chrome', 'firfox', 'safari'].some(item => item === Sniffer.browser) && MSE.isSupported(MIME_TYPE)
   }
-  if (util.typeOf(url) === 'String') {
-    mainURL = url
-  } else if (util.typeOf(url) === 'Array' && url.length) {
-    mainURL = url[0].src
-    backupURL = url[1].src
+
+  constructor (options) {
+    super(options)
+    this.mp4 = null
+    this.mse = null
+    this.waiterTimer = null
+    this.name = ''
+    this.vid = ''
+    // this.attachEvents();
   }
-  player.config._mainURL = mainURL
-  player.config._backupURL = backupURL
-  let loadData = (i = 0, time = player.currentTime, order = null, nextOrder = null) => {
+
+  afterCreate () {
+    console.log('afterCreate')
+    if (!M4APlayer.isSupported()) {
+      return
+    }
+    // BasePlugin.defineGetterOrSetter(player, {
+    //   '__url': {
+    //     get: () => {
+    //       return this.mse ? this.mse.url : player.config.url
+    //     }
+    //   }
+    // });
+    this.attachEvents()
+  }
+
+  beforePlayerInit () {
+    if (!M4APlayer.isSupported()) {
+      return
+    }
+    const { playerConfig, player } = this
+    let urlInfo = {}
+    if (Util.typeOf(playerConfig.url) === 'Array') {
+      urlInfo = playerConfig.url
+    } else {
+      urlInfo = [{
+        src: playerConfig.url,
+        name: playerConfig.name,
+        vid: playerConfig.vid,
+        poster: playerConfig.poster
+      }]
+    }
+    if (!urlInfo[0].src) {
+      player.emit('error', new Errors('other', player.config.vid))
+      return
+    }
+    this.urlInfo = urlInfo
+    this.url = urlInfo[0].src
+    // if (config.segPlay) {
+    //   let sp = config.segPlay
+    //   player.cut(sp.start, sp.end).then(blob => {
+    //     if (blob) {
+    //       player.config.url = URL.createObjectURL(blob)
+    //       player.start(player.config.url)
+    //     }
+    //   }, () => {
+    //     console.log('error')
+    //   })
+    //   return
+    // }
+
+    return this.initM4a(this.url).then(result => {
+      const mp4 = result[0]; const mse = result[1]
+      player.mp4 = mp4
+      player.mse = mse
+      player.url = this.mse.url
+      mp4.on('error', err => {
+        errorHandle(player, err)
+      })
+    })
+  }
+
+  attachEvents () {
+    const { player } = this
+    // player.once('canplay', () => {
+    //   // safari decoder time offset
+    //   if (Sniffer.browser === 'safari' && player.buffered.length) {
+    //     let start = player.buffered.start(0);
+    //     player.currentTime = start + 0.1;
+    //   }
+    // });
+    player.on(Events.TIME_UPDATE, this.onTimeupdate)
+    player.on(Events.SEEKING, this.onSeeking)
+    player.on(Events.WAITING, this.onWaiting)
+    player.on(Events.REPLAY, this.replay)
+  }
+
+  detachEvents () {
+    const { player } = this
+    player.off(Events.TIME_UPDATE, this.onTimeupdate)
+    player.off(Events.SEEKING, this.onSeeking)
+    player.off(Events.WAITING, this.onWaiting)
+    player.off(Events.REPLAY, this.replay)
+  }
+
+  loadData (i = 0, time = 0, order = null, nextOrder = null) {
+    const { player } = this
     if (player.timer) {
       clearTimeout(player.timer)
     }
@@ -74,7 +155,7 @@ let m4aplayer = function () {
     player.timer = setTimeout(function () {
       player.mp4.seek(time, order, nextOrder).then(buffer => {
         if (buffer) {
-          let mse = player.mse
+          const mse = player.mse
           mse.updating = true
           mse.appendBuffer(buffer)
           mse.once('updateend', () => {
@@ -84,347 +165,233 @@ let m4aplayer = function () {
       }, () => {
         if (i < 10) {
           setTimeout(function () {
-            loadData(i + 1)
+            this.loadData(i + 1)
           }, 2000)
         }
       })
     }, 50)
   }
-  let init = (url, replaying = false) => {
-    let mp4 = new MP4(url)
-    mp4.reqTimeLength = player.config.reqTimeLength || 5
-    let mse
+
+  initM4a (url, replaying = false) {
+    const { config } = this
+    this.mp4 = new MP4(url)
+    this.mp4.reqTimeLength = config.reqTimeLength
     return new Promise((resolve, reject) => {
-      mp4.once('mdatReady', () => {
-        mse = new MSE()
+      this.mp4.once('mdatReady', () => {
+        this.mse = new MSE()
         if (replaying) {
-          mse.replaying = true
+          this.mse.replaying = true
         }
-        mse.on('sourceopen', function () {
-          mse.appendBuffer(mp4.packMeta(mp4.meta))
-          mse.once('updateend', loadData.bind(player))
+        this.mse.on('sourceopen', () => {
+          this.mse.appendBuffer(this.mp4.packMeta(this.mp4.meta))
+          this.mse.once('updateend', () => {
+            this.loadData()
+          })
         })
-        mse.on('error', function (e) {
+        this.mse.on('error', function (e) {
           reject(e)
         })
-        resolve([mp4, mse])
+        resolve([this.mp4, this.mse])
+      })
+      this.mp4.on('error', (e) => {
+        reject(e)
+      })
+    })
+  }
+
+  cut (url, start = 0, end) {
+    const segment = new Buffer()
+    return new Promise((resolve, reject) => {
+      const mp4 = new MP4(url)
+      mp4.once('mdatReady', () => {
+        if (!end || end <= start) {
+          end = start + 15
+        }
+        if (end > mp4.meta.audioDuration) {
+          start = mp4.meta.audioDuration - (end - start)
+          end = mp4.meta.audioDuration
+        }
+        mp4.reqTimeLength = end - start
+        mp4.cut = true
+        mp4.seek(start).then(buffer => {
+          if (buffer) {
+            const meta = Util.deepCopy({
+              duration: mp4.reqTimeLength,
+              audioDuration: mp4.reqTimeLength,
+              endTime: mp4.reqTimeLength
+            }, mp4.meta)
+            meta.duration = mp4.reqTimeLength
+            meta.audioDuration = mp4.reqTimeLength
+            meta.endTime = mp4.reqTimeLength
+            segment.write(mp4.packMeta(meta), buffer)
+            resolve(new window.Blob([segment.buffer], { type: MIME_TYPE }))
+          }
+        })
       })
       mp4.on('error', (e) => {
         reject(e)
       })
     })
   }
-  if (['chrome', 'firfox', 'safari'].some(item => item === sniffer.browser) && MSE.isSupported('audio/mp4; codecs="mp4a.40.5"')) {
-    let _start = player.start
-    if (!rule.call(player)) {
-      return false
-    }
-    player.cut = function (start = 0, end) {
-      let segment = new Buffer()
-      return new Promise((resolve, reject) => {
-        let mp4 = new MP4(url)
-        mp4.once('mdatReady', () => {
-          if (!end || end <= start) {
-            end = start + 15
-          }
-          if (end > mp4.meta.audioDuration) {
-            start = mp4.meta.audioDuration - (end - start)
-            end = mp4.meta.audioDuration
-          }
-          mp4.reqTimeLength = end - start
-          mp4.cut = true
-          mp4.seek(start).then(buffer => {
-            if (buffer) {
-              let meta = Player.util.deepCopy({
-                duration: mp4.reqTimeLength,
-                audioDuration: mp4.reqTimeLength,
-                endTime: mp4.reqTimeLength
-              }, mp4.meta)
-              meta.duration = mp4.reqTimeLength
-              meta.audioDuration = mp4.reqTimeLength
-              meta.endTime = mp4.reqTimeLength
-              segment.write(mp4.packMeta(meta), buffer)
-              resolve(new Blob([segment.buffer], {type: 'audio/mp4; codecs="mp4a.40.5"'}))
-            }
-          })
-        })
-        mp4.on('error', (e) => {
-          reject(e)
-        })
-      })
-    }
-    if (player.config.segPlay) {
-      let sp = player.config.segPlay
-      player.cut(sp.start, sp.end).then(blob => {
-        if (blob) {
-          player.config.url = URL.createObjectURL(blob)
-          player.start(player.config.url)
-        }
-      }, () => {
-        console.log('error')
-      })
-      return
-    }
-    Object.defineProperty(player, 'src', {
-      get () {
-        return player.currentSrc
-      },
-      set (url) {
-        player.config.url = url
-        if (!player.paused) {
-          player.pause()
-          player.once('pause', () => {
-            player.start(url)
-          })
-          player.once('canplay', () => {
-            player.play()
-          })
+
+  switchURL (url) {
+
+  }
+
+  onTimeupdate = (e) => {
+    const { config, player, mse, mp4 } = this
+    // let mse = player.mse; let mp4 = player.mp4
+    if (mse && !mse.updating && mp4.canDownload) {
+      const timeRage = mp4.timeRage
+      const range = player.getBufferedRange()
+      const cacheMaxTime = player.currentTime + config.preloadTime
+      if (range[1] - cacheMaxTime > 0) {
+        return
+      }
+      timeRage.every((item, idx) => {
+        if (range[1] === 0) {
+          this.loadData(5)
+          return false
         } else {
-          player.start(url)
-        }
-        player.once('canplay', () => {
-          player.currentTime = 0
-        })
-      },
-      configurable: true
-    })
-    player.start = function (url = mainURL) {
-      init(url).then((result) => {
-        let mp4 = result[0]; let mse = result[1]
-        _start.call(player, mse.url)
-        player.mp4 = mp4
-        player.mse = mse
-        mp4.on('error', err => {
-          errorHandle(player, err)
-        })
-      }, err => {
-        _start.call(player, url)
-        errorHandle(player, err)
-      })
-      player.once('canplay', () => {
-        // safari decoder time offset
-        if (sniffer.browser === 'safari' && player.buffered) {
-          let start = player.buffered.start(0)
-          player.currentTime = start + 0.1
-        }
-      })
-    }
-
-    player.switchURL = (url) => {
-      let mp5 = new MP4(url)
-      let mp4 = player.mp4
-      mp5.on('mdatReady', () => {
-        let timeRange = mp4.timeRage; let curTime = player.currentTime
-        timeRange = mp4.timeRage
-        let start = timeRange.find(item => item[0] - curTime > 2)[0]
-        let end = player.getBufferedRange()[1]
-        if (end - start > 0 && sniffer.browser !== 'safari') {
-          player.mse.removeBuffer(start, end)
-        }
-        player.mp4 = mp5
-        player.mse.appendBuffer(mp5.packMeta(mp5.meta))
-      })
-      mp5.on('error', err => {
-        errorHandle(player, err)
-      })
-    }
-    player.on('timeupdate', function () {
-      let mse = player.mse; let mp4 = player.mp4
-      if (mse && !mse.updating && mp4.canDownload) {
-        let timeRage = mp4.timeRage
-        let range = player.getBufferedRange(); let cacheMaxTime = player.currentTime + preloadTime
-        if (range[1] - cacheMaxTime > 0) {
-          return
-        }
-        timeRage.every((item, idx) => {
-          if (range[1] === 0) {
-            loadData(5)
-            return false
+          if (item[0].time >= range[1] && !mp4.bufferCache.has(idx)) {
+            this.loadData(0, item[0].time, item[0].order, item[1].order)
           } else {
-            if (item[0].time >= range[1] && !mp4.bufferCache.has(idx)) {
-              loadData(0, item[0].time, item[0].order, item[1].order)
-            } else {
-              return true
-            }
+            return true
           }
-        })
-        isEnded(player, mp4)// hack for older webkit
-      }
-    })
+        }
+      })
+      isEnded(player, mp4)// hack for older webkit
+    }
+  }
 
-    player.on('seeking', function () {
-      let buffered = player.buffered; let hasBuffered = false; let curTime = player.currentTime
-      Task.clear()
-      let timeRage = player.mp4.timeRage
-      if (buffered.length) {
-        for (let i = 0, len = buffered.length; i < len; i++) {
-          if (curTime >= buffered.start(i) && curTime <= buffered.end(i)) {
-            hasBuffered = true
-            break
-          }
+  onSeeking = (e) => {
+    const { player } = this
+    const buffered = player.buffered
+    let hasBuffered = false
+    const curTime = player.currentTime
+    Task.clear()
+    const timeRage = player.mp4.timeRage
+    if (buffered.length) {
+      for (let i = 0, len = buffered.length; i < len; i++) {
+        if (curTime >= buffered.start(i) && curTime <= buffered.end(i)) {
+          hasBuffered = true
+          break
         }
-        if (!hasBuffered) {
-          timeRage.every((item, idx) => {
-            if (item[0].time <= curTime && item[1].time > curTime) {
-              loadData(0, item[0].time, item[0].order, item[1].order)
-              return false
-            } else {
-              return true
-            }
-          })
-        }
-      } else {
+      }
+      if (!hasBuffered) {
         timeRage.every((item, idx) => {
           if (item[0].time <= curTime && item[1].time > curTime) {
-            loadData(0, item[0].time, item[0].order, item[1].order)
+            this.loadData(0, item[0].time, item[0].order, item[1].order)
             return false
           } else {
             return true
           }
         })
       }
-    })
+    } else {
+      timeRage.every((item, idx) => {
+        if (item[0].time <= curTime && item[1].time > curTime) {
+          this.loadData(0, item[0].time, item[0].order, item[1].order)
+          return false
+        } else {
+          return true
+        }
+      })
+    }
+  }
 
-    player.on('pause', function () {
-      Task.clear()
-    })
+  onPause = (e) => {
+    Task.clear()
+  }
 
-    player.on('playing', function () {
-      if (waiterTimer) {
-        clearTimeout(waiterTimer)
+  onPlaying = (e) => {
+    if (this.waiterTimer) {
+      clearTimeout(this.waiterTimer)
+      this.waiterTimer = null
+    }
+  }
+
+  onWaiting = (e) => {
+    const { player } = this
+    const buffered = player.buffered
+    let hasBuffered = false
+    const curTime = player.currentTime
+    Task.clear()
+    const timeRage = player.mp4.timeRage
+    if (buffered.length) {
+      for (let i = 0, len = buffered.length; i < len; i++) {
+        if (curTime >= buffered.start(i) && curTime <= buffered.end(i)) {
+          hasBuffered = true
+          break
+        }
       }
-    })
-
-    player.on('waiting', function () {
-      let buffered = player.buffered; let hasBuffered = false; let curTime = player.currentTime
-      Task.clear()
-      let timeRage = player.mp4.timeRage
-      if (buffered.length) {
-        for (let i = 0, len = buffered.length; i < len; i++) {
-          if (curTime >= buffered.start(i) && curTime <= buffered.end(i)) {
-            hasBuffered = true
-            break
-          }
-        }
-        if (!hasBuffered) {
-          timeRage.every((item, idx) => {
-            if (item[0].time <= curTime && item[1].time > curTime) {
-              loadData(0, item[0].time, item[0].order, item[1].order)
-              return false
-            } else {
-              return true
-            }
-          })
-        }
-      } else {
+      if (!hasBuffered) {
         timeRage.every((item, idx) => {
           if (item[0].time <= curTime && item[1].time > curTime) {
-            loadData(0, item[0].time, item[0].order, item[1].order)
+            this.loadData(0, item[0].time, item[0].order, item[1].order)
             return false
           } else {
             return true
           }
         })
       }
-    })
+    } else {
+      timeRage.every((item, idx) => {
+        if (item[0].time <= curTime && item[1].time > curTime) {
+          this.loadData(0, item[0].time, item[0].order, item[1].order)
+          return false
+        } else {
+          return true
+        }
+      })
+    }
+  }
 
-    player.once('destroy', () => {
-      Task.clear()
-      if (player.timer) {
-        clearTimeout(player.timer)
+  onEnded = (e) => {
+    const { player } = this
+    // player.hasEnded = true
+    if (player.config.offline) {
+      const mdatCache = new Buffer()
+      mdatCache.write(FMP4.size(player.mp4.mdatBox.size), FMP4.type('mdat'))
+      player.mp4.mdatCache.sort((a, b) => {
+        return a.start - b.start
+      })
+      let end = player.mp4.mdatCache[0].start - 1
+      player.mp4.mdatCache.forEach((item, index) => {
+        if (item.start === end + 1) {
+          mdatCache.write(item.buffer)
+          end = item.end
+        }
+      })
+      if (end !== player.mp4.mdatCache[player.mp4.mdatCache.length - 1].end) {
+        return
       }
-    })
-
-    // let playBtn = util.findDom(player.root, '.xgplayer-play');
-    // ['click', 'touchstart'].forEach(item => {
-    //   playBtn.addEventListener(item, function (e) {
-    //     e.preventDefault()
-    //     e.stopPropagation()
-    //     if (player.hasEnded) {
-    //       player.hasEnded = false
-    //       Task.clear()
-    //       player.mp4.bufferCache.clear()
-    //       // player.currentTime = 0
-    //       init(player.mp4.url, true).then((result) => {
-    //         let mp4 = result[0]; let mse = result[1]
-    //         player.src = mse.url
-    //         player.mp4 = mp4
-    //         player.mse = mse
-    //         player.mse.replaying = true
-    //         player.currentTime = 0
-    //         player.video.play().then(() => {
-    //
-    //           // player.pause()
-    //           // player.currentTime = 0
-    //         })
-    //       }, err => {
-    //         errorHandle(player, err)
-    //       })
-    //     }
-    //   })
-    // })
-
-    player.on('change', (nextItem) => {
-      player.newMusic(nextItem.src)
-      name = `${nextItem.name}`
-      vid = `${nextItem.vid}`
-      poster = `${nextItem.poster}`
-    })
-
-    player.newMusic = (url) => {
-      Task.clear()
-      player.mp4.bufferCache.clear()
-      init(url, true).then((result) => {
-        let mp4 = result[0]; let mse = result[1]
-        player.src = mse.url
-        player.mp4 = mp4
-        player.mse = mse
-        player.mse.replaying = true
-        player.currentTime = 0
+      const m4aCache = new Buffer()
+      if (player.mp4.freeBuffer) {
+        m4aCache.write(new Uint8Array(player.mp4.ftypBuffer), new Uint8Array(player.mp4.moovBuffer), new Uint8Array(player.mp4.freeBuffer), mdatCache.buffer)
+      } else {
+        m4aCache.write(new Uint8Array(player.mp4.ftypBuffer), new Uint8Array(player.mp4.moovBuffer), mdatCache.buffer)
+      }
+      const offlineVid = this.vid || this.name
+      player.database.openDB(() => {
+        player.database.addData(player.database.myDB.ojstore.name, [{ vid: offlineVid, blob: new window.Blob([m4aCache.buffer], { type: MIME_TYPE }) }])
         setTimeout(() => {
-          player.video.play()
-        }, 60)
-      }, err => {
-        errorHandle(player, err)
+          player.database.closeDB()
+        }, 5000)
       })
     }
+  }
 
-    player.on('ended', () => {
-      player.hasEnded = true
-      if (player.config.offline) {
-        let mdatCache = new Buffer()
-        mdatCache.write(FMP4.size(player.mp4.mdatBox.size), FMP4.type('mdat'))
-        player.mp4.mdatCache.sort((a, b) => {
-          return a.start - b.start
-        })
-        let end = player.mp4.mdatCache[0].start - 1
-        player.mp4.mdatCache.forEach((item, index) => {
-          if (item.start === end + 1) {
-            mdatCache.write(item.buffer)
-            end = item.end
-          }
-        })
-        if (end !== player.mp4.mdatCache[player.mp4.mdatCache.length - 1].end) {
-          return
-        }
-        let m4aCache = new Buffer()
-        if (player.mp4.freeBuffer) {
-          m4aCache.write(new Uint8Array(player.mp4.ftypBuffer), new Uint8Array(player.mp4.moovBuffer), new Uint8Array(player.mp4.freeBuffer), mdatCache.buffer)
-        } else {
-          m4aCache.write(new Uint8Array(player.mp4.ftypBuffer), new Uint8Array(player.mp4.moovBuffer), mdatCache.buffer)
-        }
-        let offlineVid = vid || name
-        if(player.database) {
-          player.database.openDB(() => {
-            player.database.addData(player.database.myDB.ojstore.name, [{vid: offlineVid, blob: new Blob([m4aCache.buffer], {type: 'audio/mp4; codecs="mp4a.40.5"'})}])
-            setTimeout(() => {
-              player.database.closeDB()
-            }, 5000)
-          })
-        }
-      }
-    })
+  replay = () => {}
+
+  destroy () {
+    const { player } = this
+    Task.clear()
+    this.detachEvents()
+    if (player.timer) {
+      clearTimeout(player.timer)
+    }
   }
 }
 
-Player.install('m4aplayer', m4aplayer)
+export default M4APlayer
