@@ -8,7 +8,7 @@ class Task {
     this.range = range
     this.id = range.join('-')
     this.on = false
-    let xhr = new window.XMLHttpRequest()
+    const xhr = new window.XMLHttpRequest()
     xhr.target = this
     xhr.responseType = 'arraybuffer'
     xhr.open('get', url)
@@ -22,7 +22,7 @@ class Task {
       xhr.target.remove()
     }
     xhr.onerror = function (e) {
-      xhr.target.emit('error', new Errors('network', '', {line: 25, handle: '[Task] constructor', msg: e.message, url}))
+      xhr.target.emit('error', new Errors('network', '', { line: 25, handle: '[Task] constructor', msg: e.message, url }))
       xhr.target.remove()
     }
     xhr.onabort = function () {
@@ -32,6 +32,7 @@ class Task {
     Task.queue.push(this)
     this.update()
   }
+
   cancel () {
     this.xhr.abort()
   }
@@ -49,10 +50,10 @@ class Task {
   }
 
   update () {
-    let Queue = Task.queue
-    let sended = Queue.filter((item) => item.on)
-    let wait = Queue.filter(item => !item.on)
-    let max = Task.limit - sended.length
+    const Queue = Task.queue
+    const sended = Queue.filter((item) => item.on)
+    const wait = Queue.filter(item => !item.on)
+    const max = Task.limit - sended.length
     wait.forEach((item, idx) => {
       if (idx < max) {
         item.run()
