@@ -1,6 +1,8 @@
 import Stream from './stream'
 import Errors from '../error'
+
 class Box {
+  static boxParse = {}
   constructor () {
     this.headSize = 8
     this.size = 0
@@ -8,6 +10,7 @@ class Box {
     this.subBox = []
     this.start = -1
   }
+
   readHeader (stream) {
     this.start = stream.position
     this.size = stream.readUint32()
@@ -16,31 +19,33 @@ class Box {
       this.size = stream.readUint64()
     } else if (this.size === 0) {
       if (this.type !== 'mdat') {
-        throw new Errors('parse', '', {line: 19, handle: '[Box] readHeader', msg: 'parse mp4 mdat box failed'})
+        throw new Errors('parse', '', { line: 19, handle: '[Box] readHeader', msg: 'parse mp4 mdat box failed' })
       }
     }
     if (this.type === 'uuid') {
-      let uuid = []
+      const uuid = []
       for (let i = 0; i < 16; i++) {
         uuid.push(stream.readUint8())
       }
     }
   }
+
   readBody (stream) {
-    let end = this.size - stream.position + this.start
-    let type = this.type
+    const end = this.size - stream.position + this.start
+    const type = this.type
     this.data = stream.buffer.slice(stream.position, stream.position + end)
     stream.position += this.data.byteLength
     let parser
     if (Box.containerBox.find(item => item === type)) {
       parser = Box.containerParser
     } else {
-      parser = Box[type]
+      parser = Box.boxParse[type]
     }
     if (parser && parser instanceof Function) {
       parser.call(this)
     }
   }
+
   read (stream) {
     this.readHeader(stream)
     this.readBody(stream)
@@ -48,10 +53,10 @@ class Box {
 
   static containerParser () {
     let stream = new Stream(this.data)
-    let size = stream.buffer.byteLength
-    let self = this
+    const size = stream.buffer.byteLength
+    const self = this
     while (stream.position < size) {
-      let box = new Box()
+      const box = new Box()
       box.readHeader(stream)
       self.subBox.push(box)
       box.readBody(stream)
