@@ -1,7 +1,9 @@
 import Box from './box'
 import Concat from 'concat-typed-array'
 import Stream from './stream'
-import * as SubBox from './box/*.js'
+import * as BoxParse from './boxParse'
+
+Box.boxParse = BoxParse
 
 class Parse {
   constructor (buffer) {
@@ -9,17 +11,17 @@ class Parse {
     this.boxes = []
     this.nextBox = null
     this.start = 0
-    let self = this
+    const self = this
     if (self.buffer) {
       Concat(Uint8Array, self.buffer, buffer)
     } else {
       self.buffer = buffer
     }
-    let bufferLength = buffer.byteLength
+    const bufferLength = buffer.byteLength
     buffer.position = 0
-    let stream = new Stream(buffer)
+    const stream = new Stream(buffer)
     while (bufferLength - stream.position >= 8) {
-      let box = new Box()
+      const box = new Box()
       box.readHeader(stream)
       if (box.size - 8 <= (bufferLength - stream.position)) {
         box.readBody(stream)
