@@ -1,9 +1,8 @@
-import Player from 'xgplayer'
-import { version } from '../version.json'
+import { Errors } from 'xgplayer'
 
-class _Errors extends Player.Errors {
+class _Errors extends Errors {
   constructor (type, vid, errd = {}, url = '') {
-    errd.version = version
+    errd.version = __VERSION__
     super(type, vid, errd)
     this.url = url
   }
