@@ -4,8 +4,9 @@ class Buffer {
   constructor () {
     this.buffer = new Uint8Array(0)
   }
+
   write (...buffer) {
-    let self = this
+    const self = this
     buffer.forEach(item => {
       if (item) {
         self.buffer = Concat(Uint8Array, self.buffer, item)
@@ -14,12 +15,7 @@ class Buffer {
       }
     })
   }
-  static writeUint16 (value) {
-    return new Uint8Array([
-      (value >> 8) & 0xff,
-      value & 0xff
-    ])
-  }
+
   static writeUint32 (value) {
     return new Uint8Array([
       value >> 24,
