@@ -2,53 +2,33 @@ import EventEmitter from 'event-emitter'
 import Errors from '../error'
 
 class MSE {
-  constructor (codecs = 'video/mp4; codecs="avc1.64001E, mp4a.40.5"', mediaType) {
+  constructor (codecs = 'video/mp4; codecs="avc1.64001E, mp4a.40.5"') {
+    const self = this
     EventEmitter(this)
     this.codecs = codecs
-    this.mediaSource = new window.MediaSource(mediaType)
+    this.mediaSource = new window.MediaSource()
     this.url = window.URL.createObjectURL(this.mediaSource)
     this.queue = []
     this.updating = false
-    this._hasDestroyed = false;
-    this._hasEndOfStream = false;
-    this._hasEndOfStreamSuccess = false;
-    this._onSourceOpen = this._onSourceOpen.bind(this);
-    this._onMediaSourceClose = this._onMediaSourceClose.bind(this);
-    this._onSourceBufferError = this._onSourceBufferError.bind(this);
-    this._onSourceBufferUpdateEnd = this._onSourceBufferUpdateEnd.bind(this);
-    this.mediaSource.addEventListener('sourceopen', this._onSourceOpen);
-    this.mediaSource.addEventListener('sourceclose', this._onMediaSourceClose);
-  }
-
-  _onSourceOpen(){
-    let self = this;
-    self.sourceBuffer = self.mediaSource.addSourceBuffer(self.codecs);
-    self.sourceBuffer.addEventListener('error', this._onSourceBufferError);
-    self.sourceBuffer.addEventListener('updateend', this._onSourceBufferUpdateEnd);
-    self.emit('sourceopen')
-  }
-
-  _onSourceBufferError(e){
-    this.emit('error', new Errors('mse', '', {line: 16, handle: '[MSE] constructor sourceopen', msg: e.message}))
-  }
-
-  _onSourceBufferUpdateEnd(){
-    let self = this;
-    self.emit('updateend');
-    if(this._hasEndOfStream && !this._hasEndOfStreamSuccess){
-      this._endOfStream();
-      return;
-    }
-    let buffer = self.queue.shift()
-    if (buffer && self.sourceBuffer && self.sourceBuffer.updating === false && self.state === 'open') {
-      self.sourceBuffer.appendBuffer(buffer)
-    } else if(buffer) {
-      self.queue.unshift(buffer);
-    }
-  }
-
-  _onMediaSourceClose(){
-    this.emit('sourceclose')
+    this.mediaSource.addEventListener('sourceopen', function () {
+      if (self.mediaSource.sourceBuffers.length === 0) {
+        self.sourceBuffer = self.mediaSource.addSourceBuffer(self.codecs)
+        self.sourceBuffer.addEventListener('error', function (e) {
+          self.emit('error', new Errors('mse', '', { line: 16, handle: '[MSE] constructor sourceopen', msg: e.message }))
+        })
+        self.sourceBuffer.addEventListener('updateend', function (e) {
+          self.emit('updateend')
+          const buffer = self.queue.shift()
+          if (buffer && self.sourceBuffer && !self.sourceBuffer.updating && self.state === 'open') {
+            self.sourceBuffer.appendBuffer(buffer)
+          }
+        })
+        self.emit('sourceopen')
+      }
+    })
+    this.mediaSource.addEventListener('sourceclose', function () {
+      self.emit('sourceclose')
+    })
   }
 
   get state () {
@@ -64,9 +44,7 @@ class MSE {
   }
 
   appendBuffer (buffer) {
-    if(!buffer) return;
-    
-    let sourceBuffer = this.sourceBuffer
+    const sourceBuffer = this.sourceBuffer
     if (sourceBuffer && !sourceBuffer.updating && this.state === 'open') {
       sourceBuffer.appendBuffer(buffer)
       return true
@@ -77,48 +55,26 @@ class MSE {
   }
 
   removeBuffer (start, end) {
-    let sourceBuffer = this.sourceBuffer
-    if (sourceBuffer && sourceBuffer.updating === false && this.state === 'open') {
-      sourceBuffer.remove(start, end)
+    this.sourceBuffer.remove(start, end)
+  }
+
+  clearBuffer () {
+    if (this.sourceBuffer) {
+      const buffered = this.sourceBuffer.buffered
+      for (let i = 0; i < buffered.length; i++) {
+        this.sourceBuffer.remove(buffered.start(i), buffered.end(i))
+      }
     }
   }
 
   endOfStream () {
-    this._hasEndOfStream = true;
-    if (this.mediaSource.readyState === 'open') {
-      if(this.sourceBuffer && !this.sourceBuffer.updating){
-        this._hasEndOfStreamSuccess = true;
-        this._endOfStream();
-      }
-      
-    }
-  }
-
-  _endOfStream(){
-    this.queue = [];
-    if(this.mediaSource.readyState === 'open'){
+    if (this.state === 'open') {
       this.mediaSource.endOfStream()
     }
   }
 
   static isSupported (codecs) {
     return window.MediaSource && window.MediaSource.isTypeSupported(codecs)
-  }
-
-  destroy(){
-    if(this._hasDestroyed){
-      return;
-    }
-    this._hasDestroyed = true;
-    window.URL.revokeObjectURL(this.url);
-    if(this.mediaSource) {
-      this.mediaSource.removeEventListener('sourceclose', this._onMediaSourceClose);
-      this.mediaSource.removeEventListener('sourceopen', this._onSourceOpen);
-    }
-    if(this.sourceBuffer) {
-      this.sourceBuffer.removeEventListener('error', this._onSourceBufferError);
-      this.sourceBuffer.removeEventListener('updateend', this._onSourceBufferUpdateEnd);
-    }
   }
 }
 
