@@ -1,4 +1,5 @@
-let util = {}
+/* eslint-disable camelcase */
+const util = {}
 
 /**
  * [使用递归查询指定type的box]
@@ -11,7 +12,7 @@ let util = {}
 util.findBox = function (root, type, result = []) {
   if (root.type !== type) {
     if (root && root.subBox) {
-      let box = root.subBox.filter(item => item.type === type)
+      const box = root.subBox.filter(item => item.type === type)
       if (box.length) {
         box.forEach(item => result.push(item))
       } else {
@@ -26,8 +27,8 @@ util.findBox = function (root, type, result = []) {
 }
 
 util.padStart = function (str, length, pad) {
-  let charstr = String(pad); let len = length >> 0; let maxlen = Math.ceil(len / charstr.length)
-  let chars = []; let r = String(str)
+  const charstr = String(pad); const len = length >> 0; let maxlen = Math.ceil(len / charstr.length)
+  const chars = []; const r = String(str)
   while (maxlen--) {
     chars.push(charstr)
   }
@@ -40,7 +41,7 @@ util.padStart = function (str, length, pad) {
  * @return {String}       [十六进制]
  */
 util.toHex = function (...value) {
-  let hex = []
+  const hex = []
   value.forEach(item => {
     hex.push(util.padStart(Number(item).toString(16), 2, 0))
   })
@@ -64,25 +65,24 @@ util.sum = function (...rst) {
  * @param  {Number} sample_order [帧次序]
  * @return {Object}              [块的位置和当前帧的偏移数]
  */
-util.stscOffset = function (stsc, sample_order, stscObj) {
+util.stscOffset = function (stsc, sample_order) {
   let chunk_index; let samples_offset = ''
-  // let chunk_start = stsc.entries.filter((item) => {
-  //   return item.first_sample <= sample_order && sample_order < item.first_sample + item.chunk_count * item.samples_per_chunk
-  // })[0]
-  let chunk_start = stscObj[sample_order];
+  const chunk_start = stsc.entries.filter((item) => {
+    return item.first_sample <= sample_order && sample_order < item.first_sample + item.chunk_count * item.samples_per_chunk
+  })[0]
   if (!chunk_start) {
-    let last_chunk = stsc.entries.pop()
+    const last_chunk = stsc.entries.pop()
     stsc.entries.push(last_chunk)
-    let chunk_offset = Math.floor((sample_order - last_chunk.first_sample) / last_chunk.samples_per_chunk)
-    let last_chunk_index = last_chunk.first_chunk + chunk_offset
-    let last_chunk_first_sample = last_chunk.first_sample + last_chunk.samples_per_chunk * chunk_offset
+    const chunk_offset = Math.floor((sample_order - last_chunk.first_sample) / last_chunk.samples_per_chunk)
+    const last_chunk_index = last_chunk.first_chunk + chunk_offset
+    const last_chunk_first_sample = last_chunk.first_sample + last_chunk.samples_per_chunk * chunk_offset
     return {
       chunk_index: last_chunk_index,
       samples_offset: [last_chunk_first_sample, sample_order]
     }
   } else {
-    let chunk_offset = Math.floor((sample_order - chunk_start.first_sample) / chunk_start.samples_per_chunk)
-    let chunk_offset_sample = chunk_start.first_sample + chunk_offset * chunk_start.samples_per_chunk
+    const chunk_offset = Math.floor((sample_order - chunk_start.first_sample) / chunk_start.samples_per_chunk)
+    const chunk_offset_sample = chunk_start.first_sample + chunk_offset * chunk_start.samples_per_chunk
     chunk_index = chunk_start.first_chunk + chunk_offset
     samples_offset = [chunk_offset_sample, sample_order]
     return {
@@ -92,20 +92,18 @@ util.stscOffset = function (stsc, sample_order, stscObj) {
   }
 }
 
-util.seekSampleOffset = function (stsc, stco, stsz, order, mdatStart, stscObj) {
-  let chunkOffset = util.stscOffset(stsc, order + 1, stscObj)
-  let sum =  util.sum.apply(null, stsz.entries.slice(chunkOffset.samples_offset[0] - 1, chunkOffset.samples_offset[1] - 1))
-  let ss = stco.entries[chunkOffset.chunk_index - 1]
-  let result = ss + sum - mdatStart
+util.seekSampleOffset = function (stsc, stco, stsz, order, mdatStart) {
+  const chunkOffset = util.stscOffset(stsc, order + 1)
+  const result = stco.entries[chunkOffset.chunk_index - 1] + util.sum.apply(null, stsz.entries.slice(chunkOffset.samples_offset[0] - 1, chunkOffset.samples_offset[1] - 1)) - mdatStart
   if (result === undefined) {
-    throw `result=${result},stco.length=${stco.entries.length},sum=${util.sum.apply(null, stsz.entries.slice(0, order))}`
+    throw new Error(`result=${result},stco.length=${stco.entries.length},sum=${util.sum.apply(null, stsz.entries.slice(0, order))}`)
   } else if (result < 0) {
-    throw `result=${result},stco.length=${stco.entries.length},sum=${util.sum.apply(null, stsz.entries.slice(0, order))}`
+    throw new Error(`result=${result},stco.length=${stco.entries.length},sum=${util.sum.apply(null, stsz.entries.slice(0, order))}`)
   }
   return result
 }
 
-util.seekSampleTime = function (stts, cttsObj, order) {
+util.seekSampleTime = function (stts, ctts, order, timeOffset = 0) {
   let time; let duration; let count = 0; let startTime = 0; let offset = 0
   stts.entry.every(item => {
     duration = item.sampleDuration
@@ -118,15 +116,23 @@ util.seekSampleTime = function (stts, cttsObj, order) {
       return true
     }
   })
-  if (cttsObj) {
-    if(cttsObj[order]){
-      offset = cttsObj[order]
-    }
+  if (ctts) {
+    let ct = 0
+    ctts.entry.every(item => {
+      ct += item.count
+      if (order < ct) {
+        offset = item.offset
+        return false
+      } else {
+        return true
+      }
+    })
   }
   if (!time) {
     time = startTime + (order - count) * duration
   }
-  return {time, duration, offset}
+  time -= timeOffset
+  return { time, duration, offset }
 }
 
 util.seekOrderSampleByTime = function (stts, timeScale, time) {
@@ -143,138 +149,55 @@ util.seekOrderSampleByTime = function (stts, timeScale, time) {
       return true
     }
   })
-  return {order, startTime}
+  return { order, startTime }
+}
+
+util.sampleCount = function (stts) {
+  let count = 0
+  stts.forEach((item, idx) => {
+    count += item.sampleCount
+  })
+  return count
 }
 
 util.seekTrakDuration = function (trak, timeScale) {
-  let stts = util.findBox(trak, 'stts'); let duration = 0
+  const stts = util.findBox(trak, 'stts'); let duration = 0
   stts.entry.forEach(item => {
     duration += item.sampleCount * item.sampleDuration
   })
   return Number(duration / timeScale).toFixed(4)
 }
 
-util.StringToArrayBuffer = function (str) {
-  let arr = new ArrayBuffer(str.length)
-  let view = new Uint8Array(arr)
-  for (let i = 0; i < str.length; i++) {
-    view[i] = str.charCodeAt(i)
-  }
-  return arr
-}
-/**
- * Convert a hex string to a Uint8Array.
- * @param {string} str
- * @return {!Uint8Array}
- * @export
- */
-util.fromHex = function (str) {
-  let arr = new Uint8Array(str.length / 2)
-  for (let i = 0; i < str.length; i += 2) {
-    arr[i / 2] = window.parseInt(str.substr(i, 2), 16)
-  }
-  return arr
-}
-util.fromCharCode = function (array) {
-  let max = 16000
-  let ret = ''
-  for (let i = 0; i < array.length; i += max) {
-    let subArray = array.subarray(i, i + max)
-    ret += String.fromCharCode.apply(null, subArray)
-  }
-
-  return ret
-}
-util.ArrayBufferToString = function (arr) {
-  let str = ''
-  let view = new Uint8Array(arr)
-  for (let i = 0; i < view.length; i++) {
-    str += String.fromCharCode(view[i])
-  }
-  return str
-}
-util.Base64ToHex = function (str) {
-  let bin = window.atob(str.replace(/-/g, '+').replace(/_/g, '/'))
-  let res = ''
-  for (let i = 0; i < bin.length; i++) {
-    res += ('0' + bin.charCodeAt(i).toString(16)).substr(-2)
-  }
-  return res
-}
-
-/**
- * Convert a Uint8Array to a base64 string.  The output will always use the
- * alternate encoding/alphabet also known as "base64url".
- * @param {!Uint8Array} arr
- * @param {boolean=} padding If true, pad the output with equals signs.
- *   Defaults to true.
- * @return {string}
- * @export
- */
-util.toBase64 = function (arr, padding) {
-  // btoa expects a "raw string" where each character is interpreted as a byte.
-  let bytes = util.fromCharCode(arr)
-  padding = (padding === undefined) ? true : padding
-  let base64 = window.btoa(bytes).replace(/\+/g, '-').replace(/\//g, '_')
-  return padding ? base64 : base64.replace(/=*$/, '')
-}
-
-util.toUTF8 = function (str) {
-  // http://stackoverflow.com/a/13691499
-  // Converts the given string to a URI encoded string.  If a character falls
-  // in the ASCII range, it is not converted; otherwise it will be converted to
-  // a series of URI escape sequences according to UTF-8.
-  // Example: 'g#€' -> 'g#%E3%82%AC'
-  let encoded = encodeURIComponent(str)
-  // Convert each escape sequence individually into a character.  Each escape
-  // sequence is interpreted as a code-point, so if an escape sequence happens
-  // to be part of a multi-byte sequence, each byte will be converted to a
-  // single character.
-  // Example: 'g#%E3%82%AC' -> '\x67\x35\xe3\x82\xac'
-  let utf8 = unescape(encoded)
-
-  let result = new Uint8Array(utf8.length)
-  for (let i = 0; i < utf8.length; ++i) {
-    result[i] = utf8.charCodeAt(i)
-  }
-  return result.buffer
-}
-
-util.bufferToString = function (value) {
-  return ("0"+(Number(value).toString(16))).slice(-2).toUpperCase()
-}
-util.strToBuf = function (str) {
-  let buffer = []
-  for(let i = 0; i < str.length; i = i + 2) {
-    buffer.push(
-      parseInt(str[i] + str[i + 1], 16)
-    )
-  }
-  return new Uint8Array(buffer)
-}
-util.str2hex = function (str) {
-  if(str === "") {
-    return ""
-  }
-  let arr = [];
-  for(let i = 0; i < str.length; i++) {
-    arr.push(str.charCodeAt(i))
-  }
-  return arr
-}
-util.parse = function (a) {
-  if (!Array.isArray(a)) {
-    let arr = [];
-    let value = '';
-    for(let i = 0; i < a.length; i++) {
-      if (i % 2) {
-        value = a[i - 1] + a[ i ]
-        arr.push(parseInt(value, 16))
-        value = ''
-      }
+export function getResponseHeaders (xhr) {
+  const headerMap = {}
+  if (xhr instanceof window.XMLHttpRequest) {
+    try {
+      const headers = xhr.getAllResponseHeaders()
+      // Convert the header string into an array
+      // of individual headers
+      const arr = headers.trim().split(/[\r\n]+/)
+      arr.forEach(function (line) {
+        const parts = line.split(': ')
+        const header = parts.shift()
+        const value = parts.join(': ')
+        headerMap[header] = value
+      })
+    } catch (error) {
     }
-    return arr
   }
-  return a.map(item => {return parseInt(item, 16)})
+  return headerMap
 }
+
+export function getTotalFromHeader (headers) {
+  const _range = headers['content-range']
+  if (!_range) {
+    return 0
+  }
+  const arr = _range.split('/')
+  if (arr.length > 1) {
+    return parseInt(arr[1], 10)
+  }
+  return 0
+}
+
 export default util
