@@ -1,74 +1,127 @@
-import {version} from '../version.json'
+// eslint-disable-next-line no-undef
+import version from './version'
+import Util from './utils/util'
+/**
+ * @typedef { import ('./player').default } Player
+ */
+
+const ERROR_TYPE_MAP = {
+  1: 'network',
+  2: 'network',
+  3: 'decoder',
+  4: 'format'
+}
+
 const ErrorTypes = {
   network: {
-    code: 1,
-    msg: '视频下载错误',
-    remark: '只要视频下载错误就使用此类型，无论是video本身的超时还是xhr的分段请求超时或者资源不存在'
+    code: 1
   },
   mse: {
-    code: 2,
-    msg: '流追加错误',
-    remark: '追加流的时候如果类型不对、无法被正确解码则会触发此类错误'
+    code: 2
   },
   parse: {
-    code: 3,
-    msg: '解析错误',
-    remark: 'mp4、hls、flv我们都是使用js进行格式解析，如果解析失败则会触发此类错误'
+    code: 3
   },
   format: {
-    code: 4,
-    msg: '格式错误',
-    remark: '如果浏览器不支持的格式导致播放错误'
+    code: 4
   },
   decoder: {
-    code: 5,
-    msg: '解码错误',
-    remark: '浏览器解码异常会抛出此类型错误'
+    code: 5
   },
   runtime: {
-    code: 6,
-    msg: '语法错误',
-    remark: '播放器语法错误'
+    code: 6
   },
   timeout: {
-    code: 7,
-    msg: '播放超时',
-    remark: '播放过程中无法正常请求下一个分段导致播放中断'
+    code: 7
   },
   other: {
-    code: 8,
-    msg: '其他错误',
-    remark: '不可知的错误或被忽略的错误类型'
+    code: 8
   }
 }
 
+/**
+ * @typedef { {
+ *   playerVersion: string,
+ *   currentTime: number,
+ *   duration: number,
+ *   ended: boolean,
+ *   readyState: number,
+ *   networkState: number,
+ *   src: any,
+ *   errorType: string,
+ *   errorCode: number,
+ *   message: string,
+ *   mediaError?: {
+ *     code: number,
+ *     message?: string
+ *   },
+ *   originError?: any,
+ *   url?: any,
+ *   host?: string,
+ *   [propName: string]: any
+ * } } IError
+ */
+
+/**
+ * @type { IError }
+ */
 class Errors {
-  constructor (type, currentTime, duration, networkState, readyState, src, currentSrc,
-    ended, errd = {line: '', handle: '', msg: '', version: ''}, errorCode, mediaError) {
-    let r = {}
-    if (arguments.length > 1) {
-      r.playerVersion = version // 播放器版本
-      r.errorType = type
-      r.domain = document.domain // domain
-      r.duration = duration // 视频时长
-      r.currentTime = currentTime
-      r.networkState = networkState
-      r.readyState = readyState
-      r.currentSrc = currentSrc
-      r.src = src
-      r.ended = ended
-      r.errd = errd // 错误详情
-      r.ex = (ErrorTypes[type] || {}).msg // 补充信息
-      r.errorCode = errorCode
-      r.mediaError = mediaError
-    } else {
-      const arg = arguments[0]
-      Object.keys(arg).map(key => {
-        r[key] = arg[key]
+  /**
+   *
+   * @param { Player } player
+   * @param { {
+   * errorType: string,
+   * errorCode: number,
+   * errorMessage: string,
+   * originError: any,
+   * ext: { [propName: string]: any; }
+   * } } errorInfo
+   * @returns
+   */
+  constructor (player, errorInfo = { errorType: '', errorCode: 0, errorMessage: '', originError: '', ext: {}, mediaError: null }) {
+    const ERROR_TYPES = player && player.i18n ? player.i18n.ERROR_TYPES : null
+    if (player.media) {
+      const mediaError = errorInfo.mediaError ? errorInfo.mediaError : (player.media.error || {})
+      const { duration, currentTime, ended, src, currentSrc } = player
+      const { readyState, networkState } = player.media
+      const r = {
+        playerVersion: version,
+        currentTime,
+        duration,
+        ended,
+        readyState,
+        networkState,
+        src: src || currentSrc,
+        errorType: errorInfo.errorType,
+        errorCode: errorInfo.errorCode || mediaError.code,
+        message: errorInfo.errorMessage || mediaError.message,
+        mediaError,
+        originError: errorInfo.originError ? errorInfo.originError.stack : '',
+        host: Util.getHostFromUrl(src || currentSrc)
+      }
+      errorInfo.ext && Object.keys(errorInfo.ext).map(key => {
+        r[key] = errorInfo.ext[key]
       })
-      r.ex = ((arg.type && ErrorTypes[arg.type]) || {}).msg
+      return r
+    } else {
+      if (arguments.length > 1) {
+        const r = {
+          playerVersion: version,
+          domain: document.domain
+        }
+        const arr = ['errorType', 'currentTime', 'duration', 'networkState', 'readyState', 'src', 'currentSrc', 'ended', 'errd', 'errorCode', 'mediaError']
+        for (let i = 0; i < arguments.length; i++) {
+          r[arr[i]] = arguments[i]
+        }
+        r.ex = ERROR_TYPES ? (ERROR_TYPES[arguments[0]] || {}).msg : ''// 补充信息
+        return r
+      }
     }
-    return r
   }
 }
-export default Errors
+
+export {
+  Errors as default,
+  ErrorTypes,
+  ERROR_TYPE_MAP
+}
